+++ main.rb
@@ -1,4 +1,5 @@
 require './app'
+require_relative './ui_methods'
 
 def main
   puts 'Welcome To School Library App!'
@@ -13,7 +14,6 @@ def main
     6 => :list_rentals_for_person,
     7 => :quit
   }
-
   loop do
     display_menu
     choice = gets.chomp.to_i
@@ -31,43 +31,6 @@ def display_menu
   puts '5. Create a rental'
   puts '6. List rentals for a person'
   puts '7. Exit'
-end
-
-def handle_choice(choice, app, choices)
-  if choices.key?(choice)
-    method_name = choices[choice]
-    send(method_name, app)
-  else
-    puts 'Invalid choice. Please try again.'
-  end
-end
-
-def list_books(app)
-  app.list_books
-end
-
-def list_people(app)
-  app.list_people
-end
-
-def create_person(app)
-  app.create_person
-end
-
-def create_book(app)
-  app.create_book
-end
-
-def create_rental(app)
-  app.create_rental
-end
-
-def list_rentals_for_person(app)
-  app.list_rentals_for_person
-end
-
-def quit(_app)
-  puts 'Goodbye!'
 end
 
 main
